requirements.txt
# For compositor (multi-image first frame)
Pillow>=10.0.0
# For optional ElevenLabs sound effects
requests>=2.28.0

# LangGraph + LlamaIndex (workflow and retriever)
langgraph>=0.2.0
llama-index>=0.10.0
python-dotenv>=1.0.0